check_temp function added
Added a check_temp function for converting from Fahrenheit to Celsius, to see if the value added by the user is above absolute zero. If the value is lower than absolute zero, then it will return an error message. Made the history/export button appear as disabled. Made the error message blank.